+++ src/postulo/core/server_urls.py
@@ -15,6 +15,11 @@ urlpatterns = [
     path("sign-in/", views.SignInView.as_view(), name="signin"),
     path("email/", views.EmailView.as_view(), name="email"),
     path("email/test/", views.EmailTestView.as_view(), name="email_test"),
+    path(
+        "email/connection/",
+        views.EmailConnectionTestView.as_view(),
+        name="email_connection_test",
+    ),
     path("plugins/", views.PluginsView.as_view(), name="plugins"),
     path("plugins/action/", views.PluginActionView.as_view(), name="plugin_action"),
     path("plugins/policy/", views.PluginPolicyView.as_view(), name="plugin_policy"),